// lib/api/generated/schemas/adminCreateDto.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Textbooked Backend Admin API
 * Backend MVP API for textbook study planning and grading.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminCreateDtoData } from './adminCreateDtoData';

export interface AdminCreateDto {
  data: AdminCreateDtoData;
}